ban --> banHang

+++ QLNhaHang/Controllers/VanPhongsController.cs
@@ -28,6 +28,7 @@ namespace QLNhaHang.Controllers
         // GET: VanPhongs
         public ActionResult Index(int id = 0, string searchString = null, int page = 1)
         {
+           
             VanPhongVM.StrUrl = Request.Url.AbsoluteUri.ToString();
             ViewBag.searchString = searchString;
             if (id != 0)
@@ -53,7 +54,7 @@ namespace QLNhaHang.Controllers
         public ActionResult Create(string strUrl)
         {
             var user = (NhanVien)Session["UserSession"];
-            if (user.Role.Name.Equals("Users"))
+            if (user.Role.Name != "Admins")
             {
                 return View("~/Views/Shared/AccessDeny.cshtml");
             }
@@ -100,7 +101,11 @@ namespace QLNhaHang.Controllers
 
         public ActionResult Edit(string strUrl, int id)
         {
-
+            var user = (NhanVien)Session["UserSession"];
+            if (user.Role.Name != "Admins")
+            {
+                return View("~/Views/Shared/AccessDeny.cshtml");
+            }
             VanPhongVM.VanPhong = _unitOfWork.vanPhongRepository.GetById(id);
             if (VanPhongVM.VanPhong == null)
             {
@@ -131,6 +136,11 @@ namespace QLNhaHang.Controllers
         [HttpPost, ActionName("Delete")]
         public ActionResult DeletePost(string strUrl, int id)
         {
+            var user = (NhanVien)Session["UserSession"];
+            if (user.Role.Name != "Admins")
+            {
+                return View("~/Views/Shared/AccessDeny.cshtml");
+            }
             var vanPhong = _unitOfWork.vanPhongRepository.GetById(id);
             _unitOfWork.vanPhongRepository.Delete(vanPhong);
             _unitOfWork.Complete();
